requirements.txt
asyncio
pyjwt
bcrypt
customtkinter
Pillow
